requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
numpy>=1.19.0
scikit-learn>=0.24.0
